PIM-5217: Move ArchivableFileWriterArchiver to Connector
